+++ lucidoc/__init__.py
@@ -6,5 +6,5 @@ __email__ = "vreuter@virginia.edu"
 from .docparse import *
 from .docstyle import *
 from .doctags import *
-from .exceptions import lucidocError
+from .exceptions import LucidocError
 from .lucidoc import *