+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
+## [0.2] - 2019-03-20
+### Added
+- Support for specific inclusion/exclusion of objects for documentation.
+- Support specification of grouping objects into separate documentation files.
+### Changed
+- Make signature for property match that of function with empty parameter list.
+
 ## [0.1] - 2019-03-11
 - First release version